Fix build warnings, separate build directories, and allow executable tracking
- Fixed C++ uninitialized variable warnings in SFML's Glyph.hpp and the Boss/Enemy structs to silence MSVC Code Analysis rule C26495.
- Separated intermediate build directories (<IntDir>) to resolve the clean/rebuild overlap warning with 2nd_try.vcxproj.
- Updated project configuration to output the game as 'Rectangle Smash.exe'.
- Modified .gitignore to allow tracking of .exe and .dll files, as well as the Release/ folder.
- Added a new 'RUNNING THE EXECUTABLE STANDALONE' section to README.md.

+++ RectangleSmash/boss.cpp
@@ -26,7 +26,7 @@ void Boss::initBoss(sf::RenderWindow* targetWindow, int diff)
     this->bossData.stateTimer = 0.f;
     this->bossData.moveDir = sf::Vector2f(1.f, 0.f);
 
-    this->bossBeam.setSize(sf::Vector2f(20.f, this->window->getSize().y));
+    this->bossBeam.setSize(sf::Vector2f(20.f, (float)this->window->getSize().y));
     this->bossBeam.setFillColor(sf::Color(255, 0, 0, 150));
 
     // Boss sprite
